Move task helpers out of Client to stay within the class length limit
The task lifecycle API (call_tool_as_task, get_task, get_task_result,
list_tasks, cancel_task) and its era / capability helpers now live in
MCPClient::Client::TaskApi, mixed into Client like TaskSupport. No
behaviour change.
