+++ lib/mcp_client.rb
@@ -24,6 +24,7 @@ require_relative 'mcp_client/server_http'
 require_relative 'mcp_client/server_streamable_http'
 require_relative 'mcp_client/server_factory'
 require_relative 'mcp_client/client/task_support'
+require_relative 'mcp_client/client/task_api'
 require_relative 'mcp_client/client'
 require_relative 'mcp_client/version'
 require_relative 'mcp_client/config_parser'
